In inplace array definition, it is now allowed that array elements are arrays of different size. (array of arrays with implicit size is created) Added new tests. Fixed error style - first letter is lower-case.

+++ NULLC/includes/io.cpp
@@ -29,7 +29,7 @@ namespace NULLCIO
 	{
 		InitConsole();
 		DWORD written;
-		WriteFile(conStdOut, data.ptr, strlen(data.ptr), &written, NULL); 
+		WriteFile(conStdOut, data.ptr, (DWORD)strlen(data.ptr), &written, NULL); 
 	}
 
 	void WriteIntConsole(int num)
